requirements.txt
degirum
degirum-tools
paho-mqtt
kafka-python
python-dotenv